feat: API処理部を別ファイルへ切り分け，また画面表示を一部変更
